Add test snapshot script
Moving to separate script directory, because otherwise jest would not
save the correct output. This is a workaround to this, so we can still
use snapshots with shellshot in our case.
New tests require a script in ./scripts/tests/name_of_test_snapshot.sh
and this then to be executed in a test in
./tests/name_of_test.test.js.