Add split ratio option
This commit reworks the internal structure of how splitting is done.
Previously we pinned the splitting to a fifty fifty value and be done
with it, calculating offsets by constants. This is replaced with a
variable approach storing the split ratio on the container node with a
clamped float. Offsets are still character based as before